Add part4 exercise 4.21. solution

+++ part4/bloglist/controllers/blogs.js
@@ -76,6 +76,18 @@ blogsRouter.post("/", async (request, response, next) => {
 
 blogsRouter.delete("/:id", async (request, response, next) => {
   try {
+    // Check for valid token
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: "token missing or invalid" });
+    }
+
+    // Only allow user to delete a blog if they created it
+    const blog = await Blog.findById(request.params.id);
+    if (blog.user.toString() !== decodedToken.id) {
+      return response.status(401).json({ error: "wrong user" });
+    }
+
     await Blog.findByIdAndRemove(request.params.id);
     response.status(204).end();
   } catch (exception) {
